advent of code - 2021/hubspot/requirements.txt
requests==2.25.1
python-dateutil==2.8.0